Update May 8, 2021 - 12:24 AM [-07:00 MST]

+++ commands/util/menu.ts
@@ -9,8 +9,20 @@ new discord.command.CommandGroup({
 }).raw(
   {
     name: 'menu',
-    aliases: ['help', 'h', 'Menu', 'Help', 'H'],
-    description: 'Displays the Pylon Menu',
+    aliases: [
+      'help',
+      'h',
+      'Menu',
+      'Help',
+      'H',
+      'm',
+      'M',
+      'commands',
+      'Commands',
+      'cmds',
+      'cmdhelp'
+    ],
+    description: 'Pylon Help Menu',
     filters: USER_PERMS
   },
   async (msg) => {
@@ -20,10 +32,11 @@ new discord.command.CommandGroup({
     await menu.setDescription(
       `Select an option.
 1️⃣: Administration
-2️⃣: Moderation Commands
-3️⃣: Social Commands
-4️⃣: Utility Commands
+2️⃣: Moderation
+3️⃣: Social
+4️⃣: Utility
 5️⃣: Slash Commands
+6️⃣: Economy
 `
     );
 
@@ -34,6 +47,7 @@ new discord.command.CommandGroup({
     await thehelpmsg.addReaction('3️⃣');
     await thehelpmsg.addReaction('4️⃣');
     await thehelpmsg.addReaction('5️⃣');
+    await thehelpmsg.addReaction('6️⃣');
     await thehelpmsg.addReaction('❌');
 
     MSGID = thehelpmsg.id;
@@ -55,11 +69,11 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
   ) {
     const option1 = new discord.Embed();
     await option1.setColor(0x3f888f);
-    await option1.setTitle('Administration Commands');
+    await option1.setTitle('Administration');
     await option1.setDescription(
       '`kick <user.mention>`: Kicks a specified user from the guild.\n`ban <user.mention> [reason]`: Ban a specified user from the guild for a specified reason.\n`slowmode: <time> [channel.mention]`: Apply slowmode to a specified channel.\n\nFields marked with the `<>` flags are required.\nFields marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.'
     );
-    const theMsg2 = await theMsg.reply(option1);
+    const theMsg1 = await theMsg.reply(option1);
     MSGID = '';
     AUTHORID = '';
     await theMsg.delete();
@@ -71,11 +85,11 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
   ) {
     const option2 = new discord.Embed();
     await option2.setColor(0x3f888f);
-    await option2.setTitle('Moderation Commands');
+    await option2.setTitle('Moderation');
     await option2.setDescription(
-      '`announce <message>`: Send an announcement to the guild announcements channel.\n`mute <user.mention>`: Mute a specified user.\n`unmute <user.mention`: Un Mute a specified user.\n`tempmute <user> <time>`: Mute a specified user for a specified amount of time.\n\nFields marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.'
+      '`announce <message>`: Send an announcement to the guild announcements channel.\n`mute <user.mention>`: Mute a specified user.\n`unmute <user.mention`: Un Mute a specified user.\n`tempmute <user> <time>`: Mute a specified user for a specified amount of time.\n`warn <user> <reason>`: Warn a user for s specified reason.\n\nFields marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.'
     );
-    const theMsg3 = await theMsg.reply(option2);
+    const theMsg2 = await theMsg.reply(option2);
     MSGID = '';
     AUTHORID = '';
     await theMsg.delete();
@@ -87,9 +101,9 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
   ) {
     const option2 = new discord.Embed();
     await option2.setColor(0x3f888f);
-    await option2.setTitle('Social Commands');
+    await option2.setTitle('Social');
     await option2.setDescription(
-      '`reddit <subreddit>`:\n`twitter sub <channel.mention> <twitter.handle>`: Add a Twitter Feed to a specified channel. *[Requires Administrator Rank]*\n`twitter unsub <channel.mention> <twitter.handle>`: Remove a Twitter feed from a specified channel. *[Requiures Administrator Rank]*\n`twitter list`: Display a list of all currently subscribes Twitter feeds.\n\nFields marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.\n\n**Note**:\nTwitter commands require admin rank to use.'
+      "`reddit <subreddit>`: Display's a random image from a specified subreddit.\n`twitter sub <channel.mention> <twitter.handle>`: Add a Twitter Feed to a specified channel. *[Requires Administrator Rank]*\n`twitter unsub <channel.mention> <twitter.handle>`: Remove a Twitter feed from a specified channel. *[Requiures Administrator Rank]*\n`twitter list`: Display a list of all currently subscribes Twitter feeds.\n\nFields marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.\n\n**Note**:\nTwitter commands require admin rank to use."
     );
     const theMsg3 = await theMsg.reply(option2);
     MSGID = '';
@@ -103,7 +117,7 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
   ) {
     const option3 = new discord.Embed();
     await option3.setColor(0x3f888f);
-    await option3.setTitle('Utility Commands');
+    await option3.setTitle('Utility');
     await option3.setDescription(
       "`search <input>`: Search for something using the worlfram api.\nPing`: Responds with the bot's ping.\n`info [user.mention]`: Displays a user's info.\n`Menu`: Displays the Pylon help menu.\n\nFields marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command."
     );
@@ -111,7 +125,7 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
     const theMsg4 = await theMsg.reply(option3);
     MSGID = '';
     AUTHORID = '';
-    await theMsg.delete();
+    await theMsg4.delete();
   }
   if (
     theReaction.emoji.name == '5️⃣' &&
@@ -124,7 +138,23 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
     await option2.setDescription(
       "`search <query>`: Perform a search using the Wolframn API.\n\nFeilds marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.'"
     );
-    const theMsg3 = await theMsg.reply(option2);
+    const theMsg5 = await theMsg.reply(option2);
+    MSGID = '';
+    AUTHORID = '';
+    await theMsg.delete();
+  }
+  if (
+    theReaction.emoji.name == '6️⃣' &&
+    theReaction.messageId == MSGID &&
+    theReaction.member.user.id == AUTHORID
+  ) {
+    const option2 = new discord.Embed();
+    await option2.setColor(0x3f888f);
+    await option2.setTitle('Economy');
+    await option2.setDescription(
+      "`rank`: Display's the user's Rank card.\n`Top`: Display's the guild's XP leaderboard.\n\nFeilds marked with the `<>` flags are required.\nFileds marked with the `[]` flags are optional.\nDo not include the `<>` and/or `[]` flags in the command.'"
+    );
+    const theMsg6 = await theMsg.reply(option2);
     MSGID = '';
     AUTHORID = '';
     await theMsg.delete();
@@ -139,10 +169,10 @@ discord.registerEventHandler('MESSAGE_REACTION_ADD', async (theReaction) => {
     await option8.setTitle('**Canceled**');
     await option8.setDescription(`You canceled the command selection`);
 
-    const theMsg4 = await theMsg.reply(option8);
+    const theMsg7 = await theMsg.reply(option8);
     MSGID = '';
     AUTHORID = '';
     await theMsg.delete();
-    setTimeout(() => theMsg4.delete(), 5000);
+    setTimeout(() => theMsg.delete(), 5000);
   }
 });
